requirements.txt
# StarPrint Dependencies
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0

# SC Data Tools (for p4k/dcb reading)
scdatatools @ git+https://gitlab.com/scmodding/frameworks/scdatatools.git@devel

# 3D Processing & Thumbnails
trimesh>=4.0.0
numpy>=1.26.0
pillow>=10.0.0
scipy>=1.11.0
